no_std support
This allows the crate to work on platforms that do not have the standard library. The alloc library is required for `Vec`.

A new default feature "std" uses the `std::vec::Vec` reexport from the automatic prelude.

libm is required for the trig functions on no_std targets. When the standard library is available, these functions are just re-exports.

But I could see an argument to switch the behavior so that `no_std` and `libm` are only enabled by a feature. "Opt-in to no_std" instead of "opt-out of std". The opt-out is more common for crates that support both no_std and std.

Run tests with `cargo test --no-default-features` and `cargo test --all-features` to ensure full coverage.